Committing part one of inlining.  This is most of the basic terms and rewrites that will be needed.
Remaining: given a inline target (a tailCall term),
I should generate a rewrite to deal with it.  The
current scheme in this commit can deal with an
inline target, but not in a general way.
(Where by general, I mean a general tailCall term.
General purpose inlining is ways away still.)