Many of the files in this commit were simply touched becuase of minor formatting changes that I made when I was bored.  Other than that, the main changes here are:
1. Updates to the set of terms to reflect the new FIR.  This is not
complete yet.  The allocation operators need to be updated, and I
need to declare terms for a few more things.  Some of the actual
expression forms may need updating.

2. Typing rules are in the middle of being updated to reflect the
idea of of using sequents to represent a FIR program.

As of this commit, lots of documentation is probably non-existant or
innaccurate at this point, and I will be going back within the next
week to clean everything up.