***Notice***
I changed the MetaPRL formatter.  If you have problems with it let me
know and I will try to fix them.

The objective of this fix was to get indentation working again.
An unrelated change was to remove specially-coded line formatting,
since the formatter already has line zones.

On the CPS end, I added the code to convert function applications
and conditionals.