This is my attempt to add some saner documentation to M_inline. As in M_dead, there's not a whole lot to say.  We need to discuss the technical report for the M language compiler sometime.
I believe we have a bug in the rewrite to perform constant folding
for division operations.  The problem is the reduction of
meta_quot will fail is we're trying to divide by zero, which means
that we could potentially end up with a program that has
extraneous MetaInt terms.

I would fix this using a conditional rewrite (which is the only
immediately obvious thing that I can think of), except I'm under
the impression we're trying to avoid conditional rewrites (can't
remember where I heard this).  In any event, I put a BUG marker on
the rewrite for now.