Cleaned up display of sequents a bit in simple_print.
Also fixed bug that was causing core_test/test_prog1 to fail.  The problem was
that this doesn't work:
.   match explode_term t with
.       << sequent ['foo] { ... }
.               when alpha_equal foo << Foo >> ->

The last line actually needs to be:
.               when alpha_equal foo << sequent_arg{Foo} >> ->

Ah, the joys of untyped languages...

That fixed test_prog1, but test_prog2 now fails.  Beware, there's lots of grubby
debugging output in core_type_infer.ml...