This is a hacked commit, to temporarily fix the weak memo problem by using arrays instead of weak arrays.  This completely fixes the problem as far as I can tell, but of course it may greatly increase the amount of memory used by MetaPRL.
Here is my suspicion.  There are two indexes into the weak memo,
a weak_descriptor (and int), and a descriptor (the int and the value).
Assumed invariant: a weak_descriptor is live iff the descriptor is live.
The implementation guarantees that the weak_descriptor is live if the
descriptor is.  We get into trouble the other way around.  If the
descriptor goes dead, the value referred to by the weak_descriptor
will change without notice.

This seems to match the current problem.  I'll do some more investigating.