Finished the TeX formatting.
The doc/latex/theories directory is alive again, just for all-theories;
all the other subdirectories are no longer used.

To make TeX documentation, run "omake tex".  As usual, you can do this
in the theory subdirectory if you like, and it will just build the
documentation for that theory.

One issue, the all-theories only builds documentation for the theories
that are included in THEORIES.  Perhaps we should print a warning in
this case.