Force the use of omake 0.7.1, with optional dependencies.
This addresses bug #59: omake does no rescan a .ml file when a .mli file
is added.  Not closing the ticket, I have to verify it fixes the problem.

A few minor changes cleaning up OMakefiles in theories/mojave.

!!! Sorry, I cried wolf !!!  The new version of omake works just fine
on mojave.  Uh, well, it happens I was checked out on the mojave_sequents
branch on mojave, and of course that version is broken.