I rolled back a bunch of changes that were made when reformatting from the ACM style to LNCS style, bringing the M paper back to the ACM style.
Hopefully, I rolled back all the right ones (e.g only the formatting ones, but
not the textual or style fixing ones). I also rolled back most of the changes
that cut some of the text out, but some of them seemed reasonable regardless
of space constraints, so I kept the removed text out.

Still need to fix a bunch of overflows.