Fixing bug 104:
It used to be the case that we could rely on sequent bindings not clashing
with any other bindings. However, once we allow nested sequents this assumption
can no longer be enforced. I've added a explode_sequent_and_rename function
(in TermMan module) that would rename all the sequent bindings as needed
to avoid clashes.