Added rudimentary support for display in a Web Browser. To use, use the command "mpopt -browser [-port <int>]", and follow the directions.  The interface is basically the same as the normal toploop.  You type commands in the input area, and MetaPRL displays the output.
Some notes about authentication.  The interface uses
challenge/response validation based on MD5 sums.
The connection requires a password, which is stored
in the clear in a protected file on the MetaPRL host,
and optionally as a cookie in your browser.  The
password is never sent in the clear over the network.

This isn't finished yet, but is pretty usable.  I would
appreciate comments.

TODO:
   1. Some exceptions are not being caught.  In general
      exception printing should be directed to the browser.
   2. Normal output should also be directed to the browser.
   3. Somehow Mozilla is not remembering cookies, which
      means that you have to enter your password more
      often than necessary.  Undoubtably this is due
      to my lack of Javascript knowledge.  If this can be
      fixed, the challenge can be updated frequently.
   4. There are some leftover files from a frame-based
      interface.  Frames are awkward, so I'll delete these
      files after the commit.
   5. I haven't tried it on Win32.