Win32 bugfix update.   MetaPRL should now compile correctly on win32. Also, the proxyedit application compiles with SSL.
Question: to use MetaPRL on win32 from a remote server, all you need
is the proxyedit application (if you want it).  Should we put it on
some standard site like www.metaprl.org?  Or should we commit it as
part of MetaPRL CVS.  The advantage of the latter is that we can
use MetaPRL even without global Internet access.