1) Group the modules into theories. Now the filesystem structure has the following form: "/group/module/rule/nn/nn/nn/..."
Note that "old-style" locations are still valid - any attempt to go to
"/module/..." would result in getting to "/group/module/..." for the group
that the module resides in (but weird things might happen if a group name
clashes with a module name :-(, so do not do that).

The group name and description are specified using THEORYNAME and THEORYDESC
variables in [O]Makefile.

2) I've added an "end" anchor to the end of the messages frame, so hopefully
it will keep scrolling to the right place.