- Fixed a few minor bugs in handling of non-sequent contexts.
- Added an input syntax for non-sequent context terms
  ( 'V<|C;...;C|>[[t]][t;...;t], where "<|...|>" and "[...]" are optional)

Note that this is only a small part of what will be needed to fully support
non-sequent contexts. The rewriter and filter code for non-sequent contexts is
still incomplete (the biggest problem is that we currently only allow term and
int paramaters in primitive rules, while non-sequent contexts require address
parameters) and would not work.