Added explicit iform category for terms.  The syntax is:
   declare iform <term_declaration>

and the usual variants.  For the others, the iform keyword goes last
for now.

   declare type iform ...

"iform" terms are not allowed after input processing, so this is the
way to make sure your helper terms do not appear in any logical
context.

There were only a few rules in mmc that broke.