Removed the $ignore check in input term checking.  I had forgotten that the check is not being used on declarations.
It brings up another issue: input checking is not being
used on terms that are being declared (because the term doesn't
exist yet).  However, it is probably a good idea to do input
checking on all the subterms.

The other place where the input checker is bypassed is for
iforms, and in the !!!WARNING!!! section of Term_grammar.  We
should probably add checking where appropriate.