Until now Jprover used to replace all originally free variables with ground terms and than recover tham at the final stage. It also used to inject its own ground terms which are actually variables in certain sense. Aleksey complained about its interaction with nested sequents. All this is now gone, jprover injects no ground terms, all of them are now variables.
The only side effect which you might not like is that I use Lm_symbol.new_number
and during check-status this index goes beyond 900000 (may be even above a million),
I don't know if it is ever exposed to a user. If this happens I can return to
usage of local counters though it might complicate the code a bit.