This is an initial experiment in reflection syntax. The syntax is as follows, based loosely on Lisp quotations.
    $`[depth] term : quotes a term at a specific binding depth
    $,term         : cancels the quotation

Examples:

    $`[d] "apply"{e1; e2} ==
        mk_bterm{d; operator[(apply{e1; e2}):op]; [e1; e2]}

    $`[d] "lambda"{x. "apply"{$,(e1 +@ x); e2}} ==
        mk_bterm{d; operator[(lambda{x. it}):op];
           [bind{x. mk_bterm{d -@ 1; operator[(apply{it; it}):op]; [e1 +@ x; e2]}]

Don't worry that the term doesn't make sense, I'm just showing the syntax.

(I don't know if bind{x. ...} is the right thing to do on the subterm.
In fact, I'm just guessing about what reflected terms look like.)

To simplify a little, I would like to add a depth marker $#.
This would simplify the code in the case where the depth
is computed bottom-up.

(I'm using the simpler syntax here--the term is similar to the
one above).

    $`(fun x -> ($,(e1 + x) $#e2))

In this case, the depth of the lambda is (depth{e2} +@ 1).

!! Note !!

I hardcoded this iform into Filter_grammar.  I think it should go
somewhere else, preferably into one of the Itt_hoas_* files, but
this is not easy at the moment.