Added Itt_list2.map_wf4, and downgraded map_wf2 from being the default. Reproved the theorems in Itt_hoas_lang that broke.
We actually have a general problem here.
   1. In general, we should be careful not to add rules that are
      hard to use (like map_wf2) as the default.
   2. However, in some theories they are more convenient.

For now, I added a select option, so in Itt_hoas_lang, you use selT 5
(...) to get the old behavior.

I think one possible real solution would be to add a proof "mode",
which would be a string list annotation.  Rules could be labeled with
a specific mode, like this.

   interactive map_wf2 {| intro [ProofMode "reflection"] |} : ...

And the user would use a tactic to enable some modes.

   addProofModeT : string list -> tactic

The proof mode would be inherited down the proof tree.