In the generic term grammar, instead of requiring that bterms be escaped, require that tuples be enclosed in parens.
That is, instead of this term,
    "foo"{\x, y, z. x, y, z}
you write:
    "foo"{x, y, z. (x, y, z)}

You may get slightly annoyed--instead of writing <:xterm< a, b >>, you
have to write <:xterm< (a, b) >>.  However, I think the parens are good
style, and they are a simple solution to the bvars vs. tuples problem.

It *is* possible to define a grammar where you write
    "foo"{x, y, z. x, y, z}
which is compatible with the existing term grammar.  However, this
would require that we have one nonterminal for "non-variable" terms, and
another for terms with variables.  For example, when you extend the
grammar, you would have write:

    xterm_nonvar_term{'a +@ 'b} <--
       xterm_any_term{'a}; tok_plus; xterm_any_term{'b}

But I think this would be too confusing to people (plus, the core grammar
size would double).

Let me know if you prefer the most general grammar.