Updated the generic term grammar so that opnames need to be quoted less often.  Basically, if an id is followed by a left-brace, it is an opname.  Here are some examples.
   unit{}           ==  << unit >>
   lambda{x. x}     ==  << lambda{x. 'x} >>
   "unit"           ==  << unit >>
   Itt_rfun!lambda{x. x} == << Itt_rfun!lambda{x. 'x} >>

Also, first-order variables are not simple terms, so they
must be enclosed in parens for application lists.  This is
basically because using ! for first-order variables was
an unfortunate choice (we might like !foo to be the term
!"foo").

   Itt_rfun (!lambda) == << 'Itt_rfun 'lambda >>

However, opnames still need to be quoted for terms that have
parameters.  The basic problem here is that variables are not
quoted, and sovars use the same square brackets as parameters.

   number           == << 'number >>
   number[1]        == << 'number[1] >>
   "number"[1]      == << number[1:n] >>

We could potentially require that all parameter lists include
constraints.

   number[1]        == << 'number[1] >>
   number[1:n]      == << number[1:n] >>

To me, this seems a little confusing.

Alternately (and this would require a fair amount of work), we could
require braces here too.

   number[1]{}      == << number[1:n] >>
   number[1]        == << 'number[1] >>