Added the subtyping rules for FSub.
Finished rule quoting.  It includes both context and term arguments to
second-order variables and contexts, which turn into the appropriate
bind/subst forms.

Of course, this means that there are bind/subst everywhere-- who
knows if this will actually be useable.  The next step is defining the
least-fixed-point with srec, then we'll see.