One problem with using srec to define the derivations is that it requires a particular universe level.
It may be better to use the union of the finite unrollings,
so I'll try it.

Note that the type of proof rules already uses a universe level--
a proof checker has the type (ProofStep -> univ[1]).  At the
moment, I doubt this needs to be polymorphic.  However, it is
possible to imagine some very strange logics with larger types...